Move inputWrap component and Update imports across files
