+++ src/components/dashboard/postproperty/pages/location/HoodAddress.tsx
@@ -2,9 +2,12 @@ import { ReactNode, useState } from 'react';
 import OptGroup from '~/components/herosection/Optgroup';
 import FormControl from '~/components/reusable/FormControl';
 import Address from '~/components/reusable/placesAutocomplete/Address';
-import InputWrap from '~/components/dashboard/postproperty/pages/miscellenous/InputWrap';
+import InputWrap from '~/components/dashboard/reusables/InputWrap';
 import styles from '~/components/dashboard/postproperty/pages/miscellenous/post.module.css';
-import { citiesInNigeria, statesInNigeria } from '~/components/dashboard/postproperty/pages/miscellenous//mapProps';
+import {
+  citiesInNigeria,
+  statesInNigeria,
+} from '~/components/dashboard/postproperty/pages/miscellenous//mapProps';
 
 type Props = {
   city: string;
